+++ src/lib.rs
@@ -222,6 +222,7 @@ fn create_collection_thread(
     let unnormalized_cpu = app_config_fields.unnormalized_cpu;
     let show_average_cpu = app_config_fields.show_average_cpu;
     let update_sleep = app_config_fields.update_rate;
+    let get_process_threads = app_config_fields.get_process_threads;
 
     thread::spawn(move || {
         let mut data_collector = collection::DataCollector::new(filters);
@@ -230,6 +231,7 @@ fn create_collection_thread(
         data_collector.set_use_current_cpu_total(use_current_cpu_total);
         data_collector.set_unnormalized_cpu(unnormalized_cpu);
         data_collector.set_show_average_cpu(show_average_cpu);
+        data_collector.set_get_process_threads(get_process_threads);
 
         data_collector.update_data();
         data_collector.data = Data::default();