+++ src/widgets/cpu_graph.rs
@@ -45,7 +45,7 @@ impl CpuWidgetTableData {
 }
 
 impl DataToCell<CpuWidgetColumn> for CpuWidgetTableData {
-    fn to_cell(
+    fn to_cell_text(
         &self, column: &CpuWidgetColumn, calculated_width: NonZeroU16,
     ) -> Option<Cow<'static, str>> {
         const CPU_TRUNCATE_BREAKPOINT: u16 = 5;
@@ -102,6 +102,7 @@ impl DataToCell<CpuWidgetColumn> for CpuWidgetTableData {
             } => match data_type {
                 CpuDataType::Avg => painter.styles.avg_cpu_colour,
                 CpuDataType::Cpu(index) => {
+                    let index = *index as usize;
                     painter.styles.cpu_colour_styles[index % painter.styles.cpu_colour_styles.len()]
                 }
             },