+++ CHANGELOG.md
@@ -20,6 +20,17 @@ That said, these are more guidelines rather than hardset rules, though the proje
 
 ---
 
+## [0.12.0] - Unreleased
+
+### Features
+
+- [#1793](https://github.com/ClementTsang/bottom/pull/1793): Add support for threads in Linux.
+
+### Bug Fixes
+
+- [#1800](https://github.com/ClementTsang/bottom/pull/1800): Fix colon at end of process name in Linux.
+- [#1804](https://github.com/ClementTsang/bottom/pull/1804): Draw average CPU last again.
+
 ## [0.11.1] - 2025-08-15
 
 ### Bug Fixes